requirements.txt
tensorflow
keras
mysql-connector-python
numpy
